+++ src/packages/jupyterscinco/controllers/Configure.py
@@ -1,5 +1,7 @@
 from core.BaseController import BaseController
 from utils.ConfigManager import configManager as config
+from packages.jupyterscinco.controllers.Set import Set
+
 
 # NOTE This is a special controller. It cannot be accessed via the command line
 # except through `tapis package configure`. If you added your package to the
@@ -13,4 +15,16 @@ class Configure(BaseController):
     # Use this method to set configs specific to your package using the ConfigManager
     # instance(config)
     def index(self):
-        pass
+        if config.has_section(f"package.{self.get_package()}") == False:
+            config.add_section(f"package.{self.get_package()}")
+
+        set_controller = Set()
+        method_names = set_controller.get_methods()
+        
+        for method_name in method_names:
+            method = getattr(set_controller, method_name)
+            (arg_vals, kwarg_vals) = self._prompt_arg_kwarg_vals(method)
+            method(*arg_vals, **kwarg_vals)
+
+        
+